More parts into parent class. First steps on extracted_image class.
