+++ src/config.toml
@@ -5,10 +5,10 @@ path_fine_cut = "output/2_fine_cut"
 
 
 [splitter]
-min_pixel_ratio = 0.10  # Minimum ratio of height/width of the scanned image that any contour must reach. Else the contour will be omitted. Use higher values if the ratio of the picture to the scanned album page is high and lower values otherwise.
 manual_detection_threshold = -1 # Default: -1, which enables automatic search for best value. Can be set between 0 and 255 for manual threshold. This would then refer to THRESH_BINARY_INV. Depending on the type of the scanned material it might make sense to set this value manually. If set to manual, values from 180 to 210 performed best on tested data.
+min_pixel_ratio = 0.10  # Minimum ratio of height/width of the scanned image that any contour must reach. Else the contour will be omitted. Use higher values if the ratio of the picture to the scanned album page is high and lower values otherwise.
 
 [fine_cut]
 manual_detection_threshold = -1 # Default: -1, which enables automatic search for best value. Can be set between 0 and 255 for manual threshold. This would then refer to THRESH_BINARY. Depending on the type of the scanned material it might make sense to set this value manually. 
-
-extra_crop = 10  # Extra crop: After crop/rotate often a small white border remains. This removes this. If it cuts off too much of your image, adjust this.
+min_pixel_ratio = 0.90  # Minimum ratio of height/width of the scanned image that any contour must reach. Else the contour will be omitted. Use higher values if the ratio of the picture to the scanned album page is high and lower values otherwise.
+extra_crop = 10  # Extra crop: After crop/rotate often a small white border remains. This pixel width shall be cropped from the rotated image. If it cuts off too much of your images, adjust this.
